+++ src/main/java/gladiator/philosopher/common/enumtype/GenderType.java
@@ -1,0 +1,5 @@
+package gladiator.philosopher.common.enumtype;
+
+public enum GenderType {
+  MALE, FEMALE
+}
