+++ src/main/java/gladiator/philosopher/common/enumtype/UserStatus.java
@@ -1,0 +1,7 @@
+package gladiator.philosopher.common.enumtype;
+
+public enum UserStatus {
+  ACTIVATED,
+  DELETED,
+  SUSPENDED
+}
