+++ src/main/java/gladiator/philosopher/common/exception/CustomException.java
@@ -1,0 +1,12 @@
+package gladiator.philosopher.common.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class CustomException extends RuntimeException {
+
+  private final ExceptionStatus exceptionStatus;
+
+}
