+++ src/main/java/gladiator/philosopher/common/exception/ErrorDto.java
@@ -1,0 +1,11 @@
+package gladiator.philosopher.common.exception;
+
+import lombok.Data;
+
+@Data
+public class ErrorDto {
+
+  private final int statusCode;
+  private final String message;
+
+}
