+++ src/main/java/gladiator/philosopher/security/ErrorDto.java
@@ -1,0 +1,15 @@
+package gladiator.philosopher.security;
+
+import lombok.Data;
+
+@Data
+public class ErrorDto {
+
+  private int code;
+  private String message;
+
+  public ErrorDto(int code, String message) {
+    this.code = code;
+    this.message = message;
+  }
+}
