+++ src/main/java/gladiator/philosopher/security/JwtTokenProvider.java
@@ -1,0 +1,100 @@
+package gladiator.philosopher.security;
+
+import gladiator.philosopher.common.enumtype.UserType;
+import gladiator.philosopher.security.members.MemberDetailsService;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.MalformedJwtException;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.UnsupportedJwtException;
+import io.jsonwebtoken.security.Keys;
+import java.security.Key;
+import java.util.Base64;
+import java.util.Date;
+import javax.annotation.PostConstruct;
+import javax.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
+
+
+@Setter
+@Component
+@RequiredArgsConstructor
+@Slf4j
+public class JwtTokenProvider {
+
+  private final MemberDetailsService memberDetailsService;
+  public static final String AUTHORIZATION_HEADER = "Authorization";
+  public static final String AUTHORIZATION_KEY = "auth";
+  private static final String BEARER_PREFIX = "Bearer ";
+  private static final long TOKEN_TIME = 60 * 60 * 1000 * 3000L; // 임의로 x 3000함 -> 발행시간 오지게 길 듯
+
+
+  @Value("${spring.jwt.secretKey}")
+  private String secretKey;
+  private Key key;
+  private final SignatureAlgorithm signatureAlgorithm = SignatureAlgorithm.HS256;
+
+
+  @PostConstruct
+  public void init() {
+    byte[] bytes = Base64.getDecoder().decode(secretKey);
+    key = Keys.hmacShaKeyFor(bytes);
+  }
+
+
+  public String resolveToken(HttpServletRequest request) {
+    String bearerToken = request.getHeader(AUTHORIZATION_HEADER);
+    if (StringUtils.hasText(bearerToken) && bearerToken.startsWith(BEARER_PREFIX)) {
+      return bearerToken.substring(7);
+    }
+    return null;
+  }
+
+  public String createToken(String username, UserType role) {
+    Date date = new Date();
+    return BEARER_PREFIX +
+        Jwts.builder()
+            .setSubject(username)
+            .claim(AUTHORIZATION_KEY, role)
+            .setExpiration(new Date(date.getTime() + TOKEN_TIME))
+            .setIssuedAt(date)
+            .signWith(key, signatureAlgorithm)
+            .compact();
+  }
+
+  public boolean validateToken(String token) {
+    try {
+      Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
+      return true;
+    } catch (SecurityException | MalformedJwtException e) {
+      log.info("Invalid JWT signature, 유효하지 않는 JWT 서명 입니다.");
+    } catch (ExpiredJwtException e) {
+      log.info("Expired JWT token, 만료된 JWT token 입니다.");
+    } catch (UnsupportedJwtException e) {
+      log.info("Unsupported JWT token, 지원되지 않는 JWT 토큰 입니다.");
+    } catch (IllegalArgumentException e) {
+      log.info("JWT claims is empty, 잘못된 JWT 토큰 입니다.");
+    }
+    return false;
+  }
+
+
+  public Claims getUserInfoFromToken(String token) {
+    return Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token).getBody();
+  }
+
+  public Authentication createAuthentication(String username) {
+    UserDetails userDetails = memberDetailsService.loadUserByUsername(username);
+
+    return new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
+  }
+}
