+++ src/main/java/gladiator/philosopher/security/SecurityConfig.java
@@ -1,0 +1,60 @@
+package gladiator.philosopher.security;
+
+import gladiator.philosopher.security.exception.CustomAccessDeniedHandler;
+import gladiator.philosopher.security.exception.CustomAuthenticationEntryPoint;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.factory.PasswordEncoderFactories;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+@EnableWebSecurity(debug = true)
+@Configuration
+@RequiredArgsConstructor
+@EnableGlobalMethodSecurity(securedEnabled = true, prePostEnabled = true)
+public class SecurityConfig {
+
+  private final JwtTokenProvider jwtTokenProvider;
+
+  @Bean
+  public WebSecurityCustomizer webSecurityCustomizer() {
+    // h2-console 사용 및 resources 접근 허용 설정
+    return (web) -> web.ignoring()
+        .antMatchers("/api/accounts/**");
+//                .requestMatchers(PathRequest.toH2Console())
+//                .requestMatchers(PathRequest.toStaticResources().atCommonLocations());
+  }
+
+  @Bean
+  protected SecurityFilterChain configure(HttpSecurity http) throws Exception {
+    http
+        .httpBasic().disable()
+        .csrf().disable()
+        .formLogin().disable()
+        .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+        .and()
+        .authorizeRequests()
+        .antMatchers("/api/accounts/**").permitAll()
+        .and()
+        .exceptionHandling().authenticationEntryPoint(new CustomAuthenticationEntryPoint())
+        .and()
+        .exceptionHandling().accessDeniedHandler(new CustomAccessDeniedHandler())
+        .and()
+        .addFilterBefore(new JwtAuthenticationFilter(jwtTokenProvider),
+            UsernamePasswordAuthenticationFilter.class);
+    return http.build();
+  }
+
+  @Bean
+  public static PasswordEncoder passwordEncoder() {
+    return PasswordEncoderFactories.createDelegatingPasswordEncoder();
+  }
+
+}
