+++ src/main/java/gladiator/philosopher/security/members/MemberDetails.java
@@ -1,0 +1,64 @@
+package gladiator.philosopher.security.members;
+
+import gladiator.philosopher.account.entity.Account;
+import gladiator.philosopher.common.enumtype.UserType;
+import java.util.ArrayList;
+import java.util.Collection;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+public class MemberDetails implements UserDetails {
+
+  private String email;
+  private final Account account;
+
+  public Account getMember() {
+    return account;
+  }
+
+  public MemberDetails(Account account) {
+    this.email = account.getEmail();
+    this.account = account;
+  }
+
+  @Override
+  public Collection<? extends GrantedAuthority> getAuthorities() {
+    UserType userType = account.getType();
+    String authority = userType.getAuthority();
+    SimpleGrantedAuthority simpleGrantedAuthority = new SimpleGrantedAuthority(authority);
+    Collection<GrantedAuthority> authorities = new ArrayList<>();
+    authorities.add(simpleGrantedAuthority);
+    return authorities;
+  }
+
+  @Override
+  public String getPassword() {
+    return null;
+  }
+
+  @Override
+  public String getUsername() {
+    return email;
+  }
+
+  @Override
+  public boolean isAccountNonExpired() {
+    return true;
+  }
+
+  @Override
+  public boolean isAccountNonLocked() {
+    return true;
+  }
+
+  @Override
+  public boolean isCredentialsNonExpired() {
+    return true;
+  }
+
+  @Override
+  public boolean isEnabled() {
+    return true;
+  }
+}
