+++ src/main/java/gladiator/philosopher/security/members/MemberDetailsService.java
@@ -1,0 +1,26 @@
+package gladiator.philosopher.security.members;
+
+import gladiator.philosopher.account.entity.Account;
+import gladiator.philosopher.account.repository.AccountRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class MemberDetailsService implements UserDetailsService {
+
+  private final AccountRepository accountRepository;
+
+  @Override
+  public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+    Account account = accountRepository.findByEmail(email)
+        .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
+    return new MemberDetails(account);
+  }
+
+}
