--- src/main/java/gladiator/philosopher/temp/account/entity/Account.java
@@ -1,55 +1,0 @@
-package gladiator.philosopher.temp.account.entity;
-
-import gladiator.philosopher.temp.common.BaseEntity;
-import gladiator.philosopher.temp.common.enumtype.GenderType;
-import gladiator.philosopher.temp.common.enumtype.UserStatus;
-import gladiator.philosopher.temp.common.enumtype.UserType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@Builder
-@AllArgsConstructor
-public class Account extends BaseEntity {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "account_id")
-  private Long id;
-
-  @Column(nullable = false, unique = true)
-  private String email;
-
-  @Column(nullable = false)
-  private String password;
-
-  @Column(nullable = false)
-  private Long age;
-
-  @Column(nullable = false)
-  private String nickName;
-
-  @Enumerated(EnumType.STRING)
-  @Column(nullable = false)
-  private GenderType gender;
-
-  @Enumerated(EnumType.STRING)
-  @Column(nullable = false)
-  private UserType type;
-
-  @Enumerated(EnumType.STRING)
-  @Column(nullable = false)
-  private UserStatus status;
-
-}
