--- src/main/java/gladiator/philosopher/temp/common/BaseEntity.java
@@ -1,24 +1,0 @@
-package gladiator.philosopher.temp.common;
-
-import java.time.LocalDateTime;
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-@Getter
-public abstract class BaseEntity {
-
-  @CreatedDate
-  @Column(updatable = false)
-  private LocalDateTime createDate;
-
-  @LastModifiedDate
-  private LocalDateTime modDate;
-
-}
