--- src/main/java/gladiator/philosopher/temp/common/enumtype/GenderType.java
@@ -1,5 +1,0 @@
-package gladiator.philosopher.temp.common.enumtype;
-
-public enum GenderType {
-  MALE,FEMALE
-}
