--- src/main/java/gladiator/philosopher/temp/common/enumtype/UserStatus.java
@@ -1,7 +1,0 @@
-package gladiator.philosopher.temp.common.enumtype;
-
-public enum UserStatus {
-  ACTIVATED,
-  DELETED,
-  SUSPENDED
-}
