--- src/main/java/gladiator/philosopher/temp/common/exception/CustomException.java
@@ -1,12 +1,0 @@
-package gladiator.philosopher.temp.common.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class CustomException extends RuntimeException {
-
-  private final ExceptionStatus exceptionStatus;
-
-}
