--- src/main/java/gladiator/philosopher/temp/common/exception/ErrorDto.java
@@ -1,11 +1,0 @@
-package gladiator.philosopher.temp.common.exception;
-
-import lombok.Data;
-
-@Data
-public class ErrorDto{
-
-  private final int statusCode;
-  private final String message;
-
-}
