--- src/main/java/gladiator/philosopher/temp/security/ErrorDto.java
@@ -1,15 +1,0 @@
-package gladiator.philosopher.temp.security;
-
-import lombok.Data;
-
-@Data
-public class ErrorDto {
-
-  private int code;
-  private String message;
-
-  public ErrorDto(int code, String message) {
-    this.code = code;
-    this.message = message;
-  }
-}
