--- src/main/java/gladiator/philosopher/temp/security/JwtAuthenticationFilter.java
@@ -1,62 +1,0 @@
-package gladiator.philosopher.temp.security;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import gladiator.philosopher.temp.common.exception.CustomException;
-import gladiator.philosopher.temp.common.exception.ExceptionStatus;
-import io.jsonwebtoken.Claims;
-import java.io.IOException;
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContext;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-@Slf4j
-@RequiredArgsConstructor
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
-
-  private final JwtTokenProvider jwtTokenProvider;
-
-  @Override
-  protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
-      FilterChain filterChain) throws ServletException, IOException {
-
-    String token = jwtTokenProvider.resolveToken(request);
-
-    if (token != null) {
-      if (!jwtTokenProvider.validateToken(token)) {
-        jwtExceptionHandler(response, "Token Error", HttpStatus.UNAUTHORIZED.value());
-        return;
-      }
-      Claims info = jwtTokenProvider.getUserInfoFromToken(token);
-      setAuthentication(info.getSubject());
-    }
-    filterChain.doFilter(request, response);
-  }
-
-  public void setAuthentication(String username) {
-    SecurityContext context = SecurityContextHolder.createEmptyContext();
-    Authentication authentication = jwtTokenProvider.createAuthentication(username);
-    context.setAuthentication(authentication);
-    SecurityContextHolder.setContext(context);
-  }
-
-  public void jwtExceptionHandler(HttpServletResponse response, String msg, int statusCode) {
-    response.setStatus(statusCode);
-    response.setContentType("application/json");
-    try {
-      String json = new ObjectMapper().writeValueAsString(new ErrorDto(statusCode, msg));
-      response.getWriter().write(json);
-    } catch (Exception e) {
-      log.error(e.getMessage());
-    }
-  }
-
-
-}
