--- src/main/java/gladiator/philosopher/temp/security/exception/CustomAccessDeniedHandler.java
@@ -1,29 +1,0 @@
-package gladiator.philosopher.temp.security.exception;
-
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.json.JSONObject;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-
-@Slf4j
-public class CustomAccessDeniedHandler implements AccessDeniedHandler {
-
-  @Override
-  public void handle(HttpServletRequest request, HttpServletResponse response,
-      AccessDeniedException accessDeniedException) throws IOException, ServletException {
-    log.error("권한 없는 사용자의 접근");
-
-    response.setContentType("application/json;charset=UTF-8");
-    response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-
-    JSONObject responseJson = new JSONObject();
-    responseJson.put("message", "권한 없는 사용자의 접근입니다.");
-    responseJson.put("code", 401);
-
-    response.getWriter().print(responseJson);
-  }
-}
